requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
comm==0.1.2
cssselect==1.2.0
debugpy==1.6.6
decorator==5.1.1
dnspython==2.4.2
et-xmlfile==1.1.0
executing==1.2.0
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
ipykernel==6.21.2
ipython==8.11.0
jedi==0.18.2
jmespath==1.0.1
jupyter_client==8.0.3
jupyter_core==5.2.0
lxml==4.9.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.26.0
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
pandas==2.1.1
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.1.0
prompt-toolkit==3.0.38
protobuf==4.24.4
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
pymongo==4.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==305
pyzmq==25.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==2.0.6
w3lib==2.1.2
wcwidth==0.2.6
websockets==11.0.3
wsproto==1.2.0
